refactor(business-data-access): split oversized files by concern
business-api.service.ts (557 lines) mixed ~40 HTTP methods with ~190 lines
of pure interface/type declarations. Extract the types into
business-api.models.ts; the service now imports them and re-exports
everything via `export * from './business-api.models'` so index.ts and
every existing consumer keep working unchanged.

business-site.config.ts (536 lines) mixed the BusinessSiteConfig type
family, the ~110-line DEFAULT_BUSINESS_SITE_CONFIG constant, and the
merge/clone/normalize transform functions. Split into:
- business-site.types.ts: pure type/interface declarations, no deps
- business-site.defaults.ts: DEFAULT_BUSINESS_SITE_CONFIG, depends only
  on the types
- business-site.config.ts: the transform functions, now ~215 lines,
  re-exporting the types and default constant so its public import
  surface is unchanged

business-site-blocks.ts (487 lines) actually held two independent
concerns that never referenced each other: the block/field schema data
for the landing-page editor (BUSINESS_LANDING_PAGE_BLOCK_DEFINITIONS,
~270 lines) and the BusinessSiteConfig<->ConfigDocument conversion
functions (~140 lines). Extracted the schema data into
business-site-block-definitions.ts; business-site-blocks.ts re-exports
BUSINESS_LANDING_PAGE_BLOCK_DEFINITIONS so its import path is unchanged.

Investigated but did not further split anything else: once types were
separated out, none of the remaining pieces mix unrelated concerns.

Deliberately not adding catchError/try-catch to BusinessApiService's ~40
HTTP methods in this pass -- there's no established per-method error
handling convention to follow (business-auth.service.ts only wraps
login() in a pass-through catchError/rethrow, and the one real consumer,
BusinessSiteConfigStore, already handles errors at the call site) so
this needs its own design decision rather than a mechanical addition.
