+++ libs/business-data-access/src/lib/business-site.config.ts
@@ -1,214 +1,8 @@
-export interface BusinessThemeConfig {
-  mode: 'light' | 'dark';
-  personalityId: string;
-  primaryColor: string;
-}
+import { BusinessSiteConfig, LandingSection } from './business-site.types';
+import { DEFAULT_BUSINESS_SITE_CONFIG } from './business-site.defaults';
 
-export interface BusinessTestimonial {
-  quote: string;
-  clientName: string;
-  clientDetail: string;
-}
-
-export interface BusinessFeatures {
-  store: {
-    enabled: boolean;
-  };
-  booking: {
-    enabled: boolean;
-    allowOnlinePayment?: boolean;
-  };
-  clientTasks: {
-    enabled: boolean;
-    allowClientCompletion: boolean;
-  };
-  clientPortal: {
-    enabled: boolean;
-  };
-  invoices: {
-    enabled: boolean;
-  };
-  testimonials: {
-    enabled: boolean;
-  };
-}
-
-export interface BusinessService {
-  id: string;
-  name: string;
-  description: string;
-  duration: number;
-  price: number;
-  allowOnlineBooking: boolean;
-}
-
-export interface BusinessServiceCatalogConfig {
-  source: 'manual' | 'store';
-}
-
-export interface BusinessLeadContext {
-  profileId: string;
-  appScope: string;
-}
-
-export interface BusinessSiteMetadata {
-  slug: string;
-  ownerProfileId: string;
-  ownerUserId?: string;
-  status: 'draft' | 'published';
-  onboardingCompletedAt: string;
-}
-
-export type SplitLayoutSlot = 'primary' | 'secondary';
-export type GridLayoutSlot =
-  | 'hero-wide'
-  | 'top-left'
-  | 'top-right'
-  | 'bottom-left'
-  | 'bottom-right';
-
-export type LandingSectionType =
-  | 'hero'
-  | 'about'
-  | 'services'
-  | 'store'
-  | 'testimonials'
-  | 'contact'
-  | 'booking'
-  | 'custom'
-  | 'image'
-  | 'gallery';
-
-export type LandingSectionMediaSourceType = 'url' | 'asset';
-export type LandingSectionMediaAspect =
-  | 'landscape'
-  | 'square'
-  | 'portrait'
-  | 'auto';
-export type LandingSectionMediaFit = 'cover' | 'contain';
-export type LandingSectionMediaFocalPoint =
-  | 'center'
-  | 'top'
-  | 'right'
-  | 'bottom'
-  | 'left';
-export type LandingSectionGalleryStyle = 'grid' | 'masonry';
-export type LandingSectionMotionKind =
-  | 'none'
-  | 'particle-veil'
-  | 'parallax-grid-warp'
-  | 'aurora-ribbon'
-  | 'glass-fog'
-  | 'pulse-rings'
-  | 'signal-mesh'
-  | 'topographic-drift'
-  | 'shimmer-beam';
-
-export interface LandingSectionMediaItem {
-  sourceType: LandingSectionMediaSourceType;
-  src: string;
-  alt: string;
-  caption?: string;
-  aspect?: LandingSectionMediaAspect;
-  fit?: LandingSectionMediaFit;
-  focalPoint?: LandingSectionMediaFocalPoint;
-}
-
-export interface LandingSectionGalleryConfig {
-  style?: LandingSectionGalleryStyle;
-  columns?: 2 | 3 | 4;
-  items: LandingSectionMediaItem[];
-}
-
-export interface LandingSectionMotionConfig {
-  kind?: LandingSectionMotionKind;
-  density?: number;
-  speed?: number;
-  intensity?: number;
-  height?: string;
-  reducedMotion?: boolean;
-  direction?: 'diagonal' | 'horizontal';
-  ringCount?: number;
-}
-
-export interface LandingSectionRichContentComponent {
-  instanceId: string;
-  componentType: string;
-  componentData: Record<string, unknown>;
-  position?: number;
-}
-
-export interface LandingSectionRichContent {
-  title?: string;
-  content: string;
-  injectedComponents?: LandingSectionRichContentComponent[];
-  themeConfig?: {
-    theme?: 'light' | 'dark';
-    accentColor?: string;
-  };
-}
-
-export interface LandingSection {
-  id: string;
-  type: LandingSectionType;
-  title: string;
-  enabled: boolean;
-  order: number;
-  layoutPlacement?: {
-    split?: SplitLayoutSlot;
-    grid?: GridLayoutSlot;
-  };
-  body?: string;
-  ctaLabel?: string;
-  ctaHref?: string;
-  image?: LandingSectionMediaItem;
-  gallery?: LandingSectionGalleryConfig;
-  motion?: LandingSectionMotionConfig;
-  richContent?: LandingSectionRichContent;
-}
-
-export interface BusinessSiteConfig {
-  businessType:
-    | 'fitness'
-    | 'consulting'
-    | 'coaching'
-    | 'wellness'
-    | 'general'
-    | 'accounting';
-  site: BusinessSiteMetadata;
-  leadContext: BusinessLeadContext;
-  brand: {
-    businessName: string;
-    monogram: string;
-    ownerName: string;
-    trainerName?: string;
-    tagline: string;
-    intro: string;
-    longBio: string;
-    credentials: string[];
-    specializations: string[];
-  };
-  contact: {
-    email: string;
-    phone: string;
-    location: string;
-    consultationLabel: string;
-  };
-  features: BusinessFeatures;
-  serviceCatalog: BusinessServiceCatalogConfig;
-  services: BusinessService[];
-  landingPage: {
-    sections: LandingSection[];
-    layout: 'single-column' | 'split' | 'grid';
-  };
-  clientPortal: {
-    headline: string;
-    description: string;
-    capabilities: string[];
-  };
-  testimonials: BusinessTestimonial[];
-  theme: BusinessThemeConfig;
-}
+export * from './business-site.types';
+export { DEFAULT_BUSINESS_SITE_CONFIG };
 
 function cloneSection(section: LandingSection): LandingSection {
   return {
@@ -421,116 +215,3 @@ export function cloneBusinessSiteConfig(
 ): BusinessSiteConfig {
   return mergeBusinessSiteConfig(config);
 }
-
-export const DEFAULT_BUSINESS_SITE_CONFIG: BusinessSiteConfig = {
-  businessType: 'general',
-  site: {
-    slug: 'my-business',
-    ownerProfileId: '',
-    ownerUserId: '',
-    status: 'draft',
-    onboardingCompletedAt: '',
-  },
-  leadContext: {
-    profileId: '',
-    appScope: 'business-site',
-  },
-  brand: {
-    businessName: 'My Business',
-    monogram: 'MB',
-    ownerName: 'Business Owner',
-    tagline: 'Professional services for your needs.',
-    intro: 'Quality service built around your goals.',
-    longBio: 'We provide professional services tailored to your unique needs.',
-    credentials: [],
-    specializations: [],
-  },
-  contact: {
-    email: 'hello@business.local',
-    phone: '(555) 000-0000',
-    location: 'By appointment.',
-    consultationLabel: 'Book a consultation',
-  },
-  features: {
-    store: { enabled: false },
-    booking: { enabled: true, allowOnlinePayment: false },
-    clientTasks: { enabled: false, allowClientCompletion: false },
-    clientPortal: { enabled: true },
-    invoices: { enabled: false },
-    testimonials: { enabled: true },
-  },
-  serviceCatalog: {
-    source: 'manual',
-  },
-  services: [],
-  landingPage: {
-    sections: [
-      {
-        id: 'hero',
-        type: 'hero',
-        title: 'Welcome',
-        enabled: true,
-        order: 0,
-        layoutPlacement: { split: 'primary', grid: 'hero-wide' },
-      },
-      {
-        id: 'about',
-        type: 'about',
-        title: 'About',
-        enabled: true,
-        order: 1,
-        layoutPlacement: { split: 'primary', grid: 'top-left' },
-      },
-      {
-        id: 'services',
-        type: 'services',
-        title: 'Services',
-        enabled: false,
-        order: 2,
-        layoutPlacement: { split: 'primary', grid: 'top-right' },
-      },
-      {
-        id: 'testimonials',
-        type: 'testimonials',
-        title: 'Testimonials',
-        enabled: true,
-        order: 3,
-        layoutPlacement: { split: 'secondary', grid: 'bottom-left' },
-      },
-      {
-        id: 'contact',
-        type: 'contact',
-        title: 'Contact',
-        enabled: true,
-        order: 4,
-        layoutPlacement: { split: 'secondary', grid: 'bottom-right' },
-      },
-      {
-        id: 'booking',
-        type: 'booking',
-        title: 'Book Now',
-        enabled: true,
-        order: 5,
-        layoutPlacement: { split: 'secondary', grid: 'bottom-right' },
-      },
-    ],
-    layout: 'single-column',
-  },
-  clientPortal: {
-    headline: 'The client portal keeps your plan and progress in one place.',
-    description:
-      'Clients can review sessions, follow instructions, submit updates, and stay current on billing.',
-    capabilities: [
-      'View upcoming sessions and history',
-      'Follow assigned instructions',
-      'Submit progress updates',
-      'Track invoices and packages',
-    ],
-  },
-  testimonials: [],
-  theme: {
-    mode: 'light',
-    personalityId: 'professional',
-    primaryColor: '#1f7a63',
-  },
-};
